requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
python-dotenv
pyjwt
cryptography
requests
marshmallow
marshmallow-sqlalchemy
python-dateutil
uuid
mysql-connector-python
pymysql
pytz
dependency-injector
pytest
python-logging-json
